+++ subspace/crates/sp-consensus-subspace/Cargo.toml
@@ -29,9 +29,9 @@ sp-runtime.workspace = true
 sp-runtime-interface.workspace = true
 sp-std.workspace = true
 sp-timestamp.workspace = true
-subspace-core-primitives.workspace = true
+subspace-core-primitives = { workspace = true, features = ["alloc", "scale-codec"] }
 subspace-kzg = { workspace = true, optional = true }
-subspace-verification.workspace = true
+subspace-verification = { workspace = true, features = ["alloc", "scale-codec"] }
 thiserror.workspace = true
 
 [features]
