+++ subspace/crates/subspace-archiving/Cargo.toml
@@ -20,7 +20,7 @@ bench = false
 parity-scale-codec = { workspace = true, features = ["derive"] }
 rayon = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"], optional = true }
-subspace-core-primitives.workspace = true
+subspace-core-primitives = { workspace = true, features = ["alloc", "scale-codec"] }
 subspace-erasure-coding.workspace = true
 subspace-kzg.workspace = true
 thiserror.workspace = true
