+++ subspace/crates/subspace-core-primitives/src/pieces.rs
@@ -1,34 +1,42 @@
 //! Pieces-related data structures.
 
-#[cfg(not(feature = "std"))]
-extern crate alloc;
+#[cfg(feature = "alloc")]
+mod cow_bytes;
+#[cfg(feature = "alloc")]
+mod flat_pieces;
+#[cfg(feature = "alloc")]
+mod piece;
 
-use crate::segments::{ArchivedHistorySegment, RecordedHistorySegment, SegmentIndex};
+#[cfg(feature = "alloc")]
+pub use crate::pieces::flat_pieces::FlatPieces;
+#[cfg(feature = "alloc")]
+pub use crate::pieces::piece::Piece;
+#[cfg(feature = "alloc")]
+pub use crate::segments::ArchivedHistorySegment;
+use crate::segments::{RecordedHistorySegment, SegmentIndex};
 use crate::ScalarBytes;
 #[cfg(feature = "serde")]
 use ::serde::{Deserialize, Serialize};
 #[cfg(feature = "serde")]
 use ::serde::{Deserializer, Serializer};
-#[cfg(not(feature = "std"))]
+#[cfg(feature = "alloc")]
 use alloc::boxed::Box;
-#[cfg(not(feature = "std"))]
-use alloc::format;
-#[cfg(not(feature = "std"))]
+#[cfg(feature = "alloc")]
 use alloc::vec::Vec;
-use bytes::{Bytes, BytesMut};
 use core::array::TryFromSliceError;
-use core::hash::{Hash, Hasher};
+use core::hash::Hash;
 use core::iter::Step;
-use core::{fmt, mem, slice};
+#[cfg(feature = "alloc")]
+use core::slice;
+use core::{fmt, mem};
 use derive_more::{
     Add, AddAssign, AsMut, AsRef, Deref, DerefMut, Display, Div, DivAssign, From, Into, Mul,
     MulAssign, Sub, SubAssign,
 };
-use parity_scale_codec::{Decode, Encode, EncodeLike, Input, MaxEncodedLen, Output};
-#[cfg(feature = "parallel")]
-use rayon::prelude::*;
-use scale_info::build::Fields;
-use scale_info::{Path, Type, TypeInfo};
+#[cfg(feature = "scale-codec")]
+use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
+#[cfg(feature = "scale-codec")]
+use scale_info::TypeInfo;
 #[cfg(feature = "serde")]
 use serde_big_array::BigArray;
 
@@ -44,8 +52,6 @@ use serde_big_array::BigArray;
     Eq,
     PartialEq,
     Hash,
-    Encode,
-    Decode,
     Add,
     AddAssign,
     Sub,
@@ -54,8 +60,10 @@ use serde_big_array::BigArray;
     MulAssign,
     Div,
     DivAssign,
-    TypeInfo,
-    MaxEncodedLen,
+)]
+#[cfg_attr(
+    feature = "scale-codec",
+    derive(Encode, Decode, TypeInfo, MaxEncodedLen)
 )]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[repr(transparent)]
@@ -121,14 +129,14 @@ impl PieceIndex {
     /// Segment index piece index corresponds to
     #[inline]
     pub const fn segment_index(&self) -> SegmentIndex {
-        SegmentIndex::new(self.0 / ArchivedHistorySegment::NUM_PIECES as u64)
+        SegmentIndex::new(self.0 / RecordedHistorySegment::NUM_PIECES as u64)
     }
 
     /// Position of a piece in a segment
     #[inline]
     pub const fn position(&self) -> u32 {
         // Position is statically guaranteed to fit into u32
-        (self.0 % ArchivedHistorySegment::NUM_PIECES as u64) as u32
+        (self.0 % RecordedHistorySegment::NUM_PIECES as u64) as u32
     }
 
     /// Position of a source piece in the source pieces for a segment.
@@ -187,8 +195,6 @@ impl PieceIndex {
     Eq,
     PartialEq,
     Hash,
-    Encode,
-    Decode,
     Add,
     AddAssign,
     Sub,
@@ -197,8 +203,10 @@ impl PieceIndex {
     MulAssign,
     Div,
     DivAssign,
-    TypeInfo,
-    MaxEncodedLen,
+)]
+#[cfg_attr(
+    feature = "scale-codec",
+    derive(Encode, Decode, MaxEncodedLen, TypeInfo)
 )]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[repr(transparent)]
@@ -278,7 +286,10 @@ pub struct RawRecord([[u8; ScalarBytes::SAFE_BYTES]; Self::NUM_CHUNKS]);
 
 impl fmt::Debug for RawRecord {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self.0.as_flattened()))
+        for byte in self.0.as_flattened() {
+            write!(f, "{byte:02x}")?;
+        }
+        Ok(())
     }
 }
 
@@ -382,6 +393,7 @@ impl RawRecord {
     pub const SIZE: usize = ScalarBytes::SAFE_BYTES * Self::NUM_CHUNKS;
 
     /// Create boxed value without hitting stack overflow
+    #[cfg(feature = "alloc")]
     #[inline]
     pub fn new_boxed() -> Box<Self> {
         // TODO: Should have been just `::new()`, but https://github.com/rust-lang/rust/issues/53827
@@ -439,7 +451,10 @@ pub struct Record([[u8; ScalarBytes::FULL_BYTES]; Self::NUM_CHUNKS]);
 
 impl fmt::Debug for Record {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self.0.as_flattened()))
+        for byte in self.0.as_flattened() {
+            write!(f, "{byte:02x}")?;
+        }
+        Ok(())
     }
 }
 
@@ -547,6 +562,7 @@ impl Record {
 
     /// Create boxed value without hitting stack overflow
     #[inline]
+    #[cfg(feature = "alloc")]
     pub fn new_boxed() -> Box<Self> {
         // TODO: Should have been just `::new()`, but https://github.com/rust-lang/rust/issues/53827
         // SAFETY: Data structure filled with zeroes is a valid invariant
@@ -555,6 +571,7 @@ impl Record {
 
     /// Create vector filled with zeroe records without hitting stack overflow
     #[inline]
+    #[cfg(feature = "alloc")]
     pub fn new_zero_vec(length: usize) -> Vec<Self> {
         // TODO: Should have been just `::new()`, but https://github.com/rust-lang/rust/issues/53827
         let mut records = Vec::with_capacity(length);
@@ -642,26 +659,19 @@ impl Record {
 }
 
 /// Record commitment contained within a piece.
-#[derive(
-    Copy,
-    Clone,
-    Eq,
-    PartialEq,
-    Hash,
-    Deref,
-    DerefMut,
-    From,
-    Into,
-    Encode,
-    Decode,
-    TypeInfo,
-    MaxEncodedLen,
+#[derive(Copy, Clone, Eq, PartialEq, Hash, Deref, DerefMut, From, Into)]
+#[cfg_attr(
+    feature = "scale-codec",
+    derive(Encode, Decode, TypeInfo, MaxEncodedLen)
 )]
 pub struct RecordCommitment([u8; RecordCommitment::SIZE]);
 
 impl fmt::Debug for RecordCommitment {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self.0))
+        for byte in self.0 {
+            write!(f, "{byte:02x}")?;
+        }
+        Ok(())
     }
 }
 
@@ -777,26 +787,19 @@ impl RecordCommitment {
 }
 
 /// Record witness contained within a piece.
-#[derive(
-    Copy,
-    Clone,
-    Eq,
-    PartialEq,
-    Hash,
-    Deref,
-    DerefMut,
-    From,
-    Into,
-    Encode,
-    Decode,
-    TypeInfo,
-    MaxEncodedLen,
+#[derive(Copy, Clone, Eq, PartialEq, Hash, Deref, DerefMut, From, Into)]
+#[cfg_attr(
+    feature = "scale-codec",
+    derive(Encode, Decode, TypeInfo, MaxEncodedLen)
 )]
 pub struct RecordWitness([u8; RecordWitness::SIZE]);
 
 impl fmt::Debug for RecordWitness {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self.0))
+        for byte in self.0 {
+            write!(f, "{byte:02x}")?;
+        }
+        Ok(())
     }
 }
 
@@ -911,307 +914,6 @@ impl RecordWitness {
     pub const SIZE: usize = 48;
 }
 
-enum CowBytes {
-    Shared(Bytes),
-    Owned(BytesMut),
-}
-
-impl fmt::Debug for CowBytes {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self.as_ref()))
-    }
-}
-
-impl PartialEq for CowBytes {
-    fn eq(&self, other: &Self) -> bool {
-        self.as_ref().eq(other.as_ref())
-    }
-}
-
-impl Eq for CowBytes {}
-
-impl Hash for CowBytes {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.as_ref().hash(state)
-    }
-}
-
-impl Clone for CowBytes {
-    fn clone(&self) -> Self {
-        match self {
-            Self::Shared(bytes) => Self::Shared(bytes.clone()),
-            // Always return shared clone
-            Self::Owned(bytes) => Self::Shared(Bytes::copy_from_slice(bytes)),
-        }
-    }
-}
-
-impl AsRef<[u8]> for CowBytes {
-    fn as_ref(&self) -> &[u8] {
-        match self {
-            CowBytes::Shared(bytes) => bytes.as_ref(),
-            CowBytes::Owned(bytes) => bytes.as_ref(),
-        }
-    }
-}
-
-impl AsMut<[u8]> for CowBytes {
-    #[inline]
-    fn as_mut(&mut self) -> &mut [u8] {
-        match self {
-            CowBytes::Shared(bytes) => {
-                *self = CowBytes::Owned(BytesMut::from(mem::take(bytes)));
-
-                let CowBytes::Owned(bytes) = self else {
-                    unreachable!("Just replaced; qed");
-                };
-
-                bytes.as_mut()
-            }
-            CowBytes::Owned(bytes) => bytes.as_mut(),
-        }
-    }
-}
-
-/// A piece of archival history in Subspace Network.
-///
-/// This version is allocated on the heap, for stack-allocated piece see [`PieceArray`].
-///
-/// Internally piece contains a record and corresponding witness that together with segment
-/// commitment of the segment this piece belongs to can be used to verify that a piece belongs to
-/// the actual archival history of the blockchain.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct Piece(CowBytes);
-
-impl Encode for Piece {
-    #[inline]
-    fn size_hint(&self) -> usize {
-        self.as_ref().size_hint()
-    }
-
-    #[inline]
-    fn encode_to<O: Output + ?Sized>(&self, output: &mut O) {
-        self.as_ref().encode_to(output)
-    }
-
-    #[inline]
-    fn encode(&self) -> Vec<u8> {
-        self.as_ref().encode()
-    }
-
-    #[inline]
-    fn using_encoded<R, F: FnOnce(&[u8]) -> R>(&self, f: F) -> R {
-        self.as_ref().using_encoded(f)
-    }
-}
-
-impl EncodeLike for Piece {}
-
-impl Decode for Piece {
-    fn decode<I: Input>(input: &mut I) -> Result<Self, parity_scale_codec::Error> {
-        let bytes =
-            Bytes::decode(input).map_err(|error| error.chain("Could not decode `Piece`"))?;
-
-        if bytes.len() != Self::SIZE {
-            return Err(
-                parity_scale_codec::Error::from("Incorrect Piece length").chain(format!(
-                    "Expected {} bytes, found {} bytes",
-                    Self::SIZE,
-                    bytes.len()
-                )),
-            );
-        }
-
-        Ok(Piece(CowBytes::Shared(bytes)))
-    }
-}
-
-impl TypeInfo for Piece {
-    type Identity = Self;
-
-    fn type_info() -> Type {
-        Type::builder()
-            .path(Path::new("Piece", module_path!()))
-            .docs(&["A piece of archival history in Subspace Network"])
-            .composite(
-                Fields::unnamed().field(|f| f.ty::<[u8; Piece::SIZE]>().type_name("PieceArray")),
-            )
-    }
-}
-
-#[cfg(feature = "serde")]
-impl Serialize for Piece {
-    #[inline]
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let bytes = match &self.0 {
-            CowBytes::Shared(bytes) => bytes.as_ref(),
-            CowBytes::Owned(bytes) => bytes.as_ref(),
-        };
-
-        if serializer.is_human_readable() {
-            hex::serde::serialize(bytes, serializer)
-        } else {
-            bytes.serialize(serializer)
-        }
-    }
-}
-
-#[cfg(feature = "serde")]
-impl<'de> Deserialize<'de> for Piece {
-    #[inline]
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let bytes = if deserializer.is_human_readable() {
-            hex::serde::deserialize::<_, Vec<u8>>(deserializer).and_then(|bytes| {
-                if bytes.len() == Piece::SIZE {
-                    Ok(Bytes::from(bytes))
-                } else {
-                    Err(serde::de::Error::invalid_length(
-                        bytes.len(),
-                        &format!("Expected {} bytes", Piece::SIZE).as_str(),
-                    ))
-                }
-            })?
-        } else {
-            Bytes::deserialize(deserializer)?
-        };
-
-        Ok(Piece(CowBytes::Shared(bytes)))
-    }
-}
-
-impl Default for Piece {
-    #[inline]
-    fn default() -> Self {
-        Self(CowBytes::Owned(BytesMut::zeroed(Self::SIZE)))
-    }
-}
-
-impl From<Piece> for Vec<u8> {
-    #[inline]
-    fn from(piece: Piece) -> Self {
-        match piece.0 {
-            CowBytes::Shared(bytes) => bytes.to_vec(),
-            CowBytes::Owned(bytes) => Vec::from(bytes),
-        }
-    }
-}
-
-impl TryFrom<&[u8]> for Piece {
-    type Error = ();
-
-    #[inline]
-    fn try_from(slice: &[u8]) -> Result<Self, Self::Error> {
-        if slice.len() != Self::SIZE {
-            return Err(());
-        }
-
-        Ok(Self(CowBytes::Shared(Bytes::copy_from_slice(slice))))
-    }
-}
-
-impl TryFrom<Vec<u8>> for Piece {
-    type Error = ();
-
-    #[inline]
-    fn try_from(vec: Vec<u8>) -> Result<Self, Self::Error> {
-        if vec.len() != Self::SIZE {
-            return Err(());
-        }
-
-        Ok(Self(CowBytes::Shared(Bytes::from(vec))))
-    }
-}
-
-impl TryFrom<Bytes> for Piece {
-    type Error = ();
-
-    #[inline]
-    fn try_from(bytes: Bytes) -> Result<Self, Self::Error> {
-        if bytes.len() != Self::SIZE {
-            return Err(());
-        }
-
-        Ok(Self(CowBytes::Shared(bytes)))
-    }
-}
-
-impl TryFrom<BytesMut> for Piece {
-    type Error = ();
-
-    #[inline]
-    fn try_from(bytes: BytesMut) -> Result<Self, Self::Error> {
-        if bytes.len() != Self::SIZE {
-            return Err(());
-        }
-
-        Ok(Self(CowBytes::Owned(bytes)))
-    }
-}
-
-impl From<&PieceArray> for Piece {
-    #[inline]
-    fn from(value: &PieceArray) -> Self {
-        Self(CowBytes::Shared(Bytes::copy_from_slice(value.as_ref())))
-    }
-}
-
-impl Deref for Piece {
-    type Target = PieceArray;
-
-    #[inline]
-    fn deref(&self) -> &Self::Target {
-        <&[u8; Self::SIZE]>::try_from(self.as_ref())
-            .expect("Slice of memory has correct length; qed")
-            .into()
-    }
-}
-
-impl DerefMut for Piece {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        <&mut [u8; Self::SIZE]>::try_from(self.as_mut())
-            .expect("Slice of memory has correct length; qed")
-            .into()
-    }
-}
-
-impl AsRef<[u8]> for Piece {
-    #[inline]
-    fn as_ref(&self) -> &[u8] {
-        self.0.as_ref()
-    }
-}
-
-impl AsMut<[u8]> for Piece {
-    #[inline]
-    fn as_mut(&mut self) -> &mut [u8] {
-        self.0.as_mut()
-    }
-}
-
-impl Piece {
-    /// Size of a piece (in bytes).
-    pub const SIZE: usize = Record::SIZE + RecordCommitment::SIZE + RecordWitness::SIZE;
-
-    /// Ensure piece contains cheaply cloneable shared data.
-    ///
-    /// Internally piece uses CoW mechanism and can store either mutable owned data or data that is
-    /// cheap to clone, calling this method will ensure further clones will not result in additional
-    /// memory allocations.
-    pub fn to_shared(self) -> Self {
-        Self(match self.0 {
-            CowBytes::Shared(bytes) => CowBytes::Shared(bytes),
-            CowBytes::Owned(bytes) => CowBytes::Shared(bytes.freeze()),
-        })
-    }
-}
-
 /// A piece of archival history in Subspace Network.
 ///
 /// This version is allocated on the stack, for heap-allocated piece see [`Piece`].
@@ -1221,18 +923,21 @@ impl Piece {
 /// the actual archival history of the blockchain.
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Deref, DerefMut, AsRef, AsMut)]
 #[repr(transparent)]
-pub struct PieceArray([u8; Piece::SIZE]);
+pub struct PieceArray([u8; Self::SIZE]);
 
 impl fmt::Debug for PieceArray {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self.0))
+        for byte in self.0 {
+            write!(f, "{byte:02x}")?;
+        }
+        Ok(())
     }
 }
 
 impl Default for PieceArray {
     #[inline]
     fn default() -> Self {
-        Self([0u8; Piece::SIZE])
+        Self([0u8; Self::SIZE])
     }
 }
 
@@ -1250,7 +955,7 @@ impl AsMut<[u8]> for PieceArray {
     }
 }
 
-impl From<&PieceArray> for &[u8; Piece::SIZE] {
+impl From<&PieceArray> for &[u8; PieceArray::SIZE] {
     #[inline]
     fn from(value: &PieceArray) -> Self {
         // SAFETY: `PieceArray` is `#[repr(transparent)]` and guaranteed to have the same memory
@@ -1259,16 +964,16 @@ impl From<&PieceArray> for &[u8; Piece::SIZE] {
     }
 }
 
-impl From<&[u8; Piece::SIZE]> for &PieceArray {
+impl From<&[u8; PieceArray::SIZE]> for &PieceArray {
     #[inline]
-    fn from(value: &[u8; Piece::SIZE]) -> Self {
+    fn from(value: &[u8; PieceArray::SIZE]) -> Self {
         // SAFETY: `PieceArray` is `#[repr(transparent)]` and guaranteed to have the same memory
         // layout
         unsafe { mem::transmute(value) }
     }
 }
 
-impl From<&mut PieceArray> for &mut [u8; Piece::SIZE] {
+impl From<&mut PieceArray> for &mut [u8; PieceArray::SIZE] {
     #[inline]
     fn from(value: &mut PieceArray) -> Self {
         // SAFETY: `PieceArray` is `#[repr(transparent)]` and guaranteed to have the same memory
@@ -1277,9 +982,9 @@ impl From<&mut PieceArray> for &mut [u8; Piece::SIZE] {
     }
 }
 
-impl From<&mut [u8; Piece::SIZE]> for &mut PieceArray {
+impl From<&mut [u8; PieceArray::SIZE]> for &mut PieceArray {
     #[inline]
-    fn from(value: &mut [u8; Piece::SIZE]) -> Self {
+    fn from(value: &mut [u8; PieceArray::SIZE]) -> Self {
         // SAFETY: `PieceArray` is `#[repr(transparent)]` and guaranteed to have the same memory
         // layout
         unsafe { mem::transmute(value) }
@@ -1287,8 +992,12 @@ impl From<&mut [u8; Piece::SIZE]> for &mut PieceArray {
 }
 
 impl PieceArray {
+    /// Size of a piece (in bytes).
+    pub const SIZE: usize = Record::SIZE + RecordCommitment::SIZE + RecordWitness::SIZE;
+
     /// Create boxed value without hitting stack overflow
     #[inline]
+    #[cfg(feature = "alloc")]
     pub fn new_boxed() -> Box<Self> {
         // TODO: Should have been just `::new()`, but https://github.com/rust-lang/rust/issues/53827
         // SAFETY: Data structure filled with zeroes is a valid invariant
@@ -1366,7 +1075,7 @@ impl PieceArray {
     /// Convenient conversion from slice of piece array to underlying representation for efficiency
     /// purposes.
     #[inline]
-    pub fn slice_to_repr(value: &[Self]) -> &[[u8; Piece::SIZE]] {
+    pub fn slice_to_repr(value: &[Self]) -> &[[u8; Self::SIZE]] {
         // SAFETY: `PieceArray` is `#[repr(transparent)]` and guaranteed to have the same memory
         // layout
         unsafe { mem::transmute(value) }
@@ -1375,7 +1084,7 @@ impl PieceArray {
     /// Convenient conversion from slice of underlying representation to piece array for efficiency
     /// purposes.
     #[inline]
-    pub fn slice_from_repr(value: &[[u8; Piece::SIZE]]) -> &[Self] {
+    pub fn slice_from_repr(value: &[[u8; Self::SIZE]]) -> &[Self] {
         // SAFETY: `PieceArray` is `#[repr(transparent)]` and guaranteed to have the same memory
         // layout
         unsafe { mem::transmute(value) }
@@ -1384,7 +1093,7 @@ impl PieceArray {
     /// Convenient conversion from mutable slice of piece array to underlying representation for
     /// efficiency purposes.
     #[inline]
-    pub fn slice_mut_to_repr(value: &mut [Self]) -> &mut [[u8; Piece::SIZE]] {
+    pub fn slice_mut_to_repr(value: &mut [Self]) -> &mut [[u8; Self::SIZE]] {
         // SAFETY: `PieceArray` is `#[repr(transparent)]` and guaranteed to have the same memory
         // layout
         unsafe { mem::transmute(value) }
@@ -1393,169 +1102,18 @@ impl PieceArray {
     /// Convenient conversion from mutable slice of underlying representation to piece array for
     /// efficiency purposes.
     #[inline]
-    pub fn slice_mut_from_repr(value: &mut [[u8; Piece::SIZE]]) -> &mut [Self] {
+    pub fn slice_mut_from_repr(value: &mut [[u8; Self::SIZE]]) -> &mut [Self] {
         // SAFETY: `PieceArray` is `#[repr(transparent)]` and guaranteed to have the same memory
         // layout
         unsafe { mem::transmute(value) }
     }
 }
 
+#[cfg(feature = "alloc")]
 impl From<Box<PieceArray>> for Vec<u8> {
     fn from(value: Box<PieceArray>) -> Self {
         let mut value = mem::ManuallyDrop::new(value);
         // SAFETY: Always contains fixed allocation of bytes
-        unsafe { Vec::from_raw_parts(value.as_mut_ptr(), Piece::SIZE, Piece::SIZE) }
-    }
-}
-
-/// Flat representation of multiple pieces concatenated for more efficient for processing
-#[derive(Clone, PartialEq, Eq)]
-pub struct FlatPieces(CowBytes);
-
-impl fmt::Debug for FlatPieces {
-    #[inline]
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("FlatPieces").finish_non_exhaustive()
-    }
-}
-
-impl Deref for FlatPieces {
-    type Target = [PieceArray];
-
-    #[inline]
-    fn deref(&self) -> &Self::Target {
-        let bytes = self.0.as_ref();
-        // SAFETY: Bytes slice has length of multiples of piece size and lifetimes of returned data
-        // are preserved
-        let pieces = unsafe {
-            slice::from_raw_parts(
-                bytes.as_ptr() as *const [u8; Piece::SIZE],
-                bytes.len() / Piece::SIZE,
-            )
-        };
-        PieceArray::slice_from_repr(pieces)
-    }
-}
-
-impl DerefMut for FlatPieces {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        let bytes = self.0.as_mut();
-        // SAFETY: Bytes slice has length of multiples of piece size and lifetimes of returned data
-        // are preserved
-        let pieces = unsafe {
-            slice::from_raw_parts_mut(
-                bytes.as_mut_ptr() as *mut [u8; Piece::SIZE],
-                bytes.len() / Piece::SIZE,
-            )
-        };
-        PieceArray::slice_mut_from_repr(pieces)
-    }
-}
-
-impl FlatPieces {
-    /// Allocate `FlatPieces` that will hold `piece_count` pieces filled with zeroes
-    #[inline]
-    pub fn new(piece_count: usize) -> Self {
-        Self(CowBytes::Owned(BytesMut::zeroed(piece_count * Piece::SIZE)))
-    }
-
-    /// Iterate over all pieces.
-    ///
-    /// NOTE: Unless [`Self::to_shared`] was called first, iterator may have to allocate each piece
-    /// from scratch, which is rarely a desired behavior.
-    #[inline]
-    pub fn pieces(&self) -> Box<dyn ExactSizeIterator<Item = Piece> + '_> {
-        match &self.0 {
-            CowBytes::Shared(bytes) => Box::new(
-                bytes
-                    .chunks_exact(Piece::SIZE)
-                    .map(|slice| Piece(CowBytes::Shared(bytes.slice_ref(slice)))),
-            ),
-            CowBytes::Owned(bytes) => Box::new(
-                bytes
-                    .chunks_exact(Piece::SIZE)
-                    .map(|slice| Piece(CowBytes::Shared(Bytes::copy_from_slice(slice)))),
-            ),
-        }
-    }
-
-    /// Iterator over source pieces (even indices)
-    #[inline]
-    pub fn source_pieces(&self) -> impl ExactSizeIterator<Item = Piece> + '_ {
-        self.pieces().step_by(2)
-    }
-
-    /// Iterator over source pieces (even indices)
-    #[inline]
-    pub fn source(&self) -> impl ExactSizeIterator<Item = &'_ PieceArray> + '_ {
-        self.iter().step_by(2)
-    }
-
-    /// Mutable iterator over source pieces (even indices)
-    #[inline]
-    pub fn source_mut(&mut self) -> impl ExactSizeIterator<Item = &'_ mut PieceArray> + '_ {
-        self.iter_mut().step_by(2)
-    }
-
-    /// Iterator over parity pieces (odd indices)
-    #[inline]
-    pub fn parity_pieces(&self) -> impl ExactSizeIterator<Item = Piece> + '_ {
-        self.pieces().skip(1).step_by(2)
-    }
-
-    /// Iterator over parity pieces (odd indices)
-    #[inline]
-    pub fn parity(&self) -> impl ExactSizeIterator<Item = &'_ PieceArray> + '_ {
-        self.iter().skip(1).step_by(2)
-    }
-
-    /// Mutable iterator over parity pieces (odd indices)
-    #[inline]
-    pub fn parity_mut(&mut self) -> impl ExactSizeIterator<Item = &'_ mut PieceArray> + '_ {
-        self.iter_mut().skip(1).step_by(2)
-    }
-
-    /// Ensure flat pieces contains cheaply cloneable shared data.
-    ///
-    /// Internally flat pieces uses CoW mechanism and can store either mutable owned data or data
-    /// that is cheap to clone, calling this method will ensure further clones and returned pieces
-    /// will not result in additional memory allocations.
-    pub fn to_shared(self) -> Self {
-        Self(match self.0 {
-            CowBytes::Shared(bytes) => CowBytes::Shared(bytes),
-            CowBytes::Owned(bytes) => CowBytes::Shared(bytes.freeze()),
-        })
-    }
-}
-
-#[cfg(feature = "parallel")]
-impl FlatPieces {
-    /// Parallel iterator over source pieces (even indices)
-    #[inline]
-    pub fn par_source(&self) -> impl IndexedParallelIterator<Item = &'_ PieceArray> + '_ {
-        self.par_iter().step_by(2)
-    }
-
-    /// Mutable parallel iterator over source pieces (even indices)
-    #[inline]
-    pub fn par_source_mut(
-        &mut self,
-    ) -> impl IndexedParallelIterator<Item = &'_ mut PieceArray> + '_ {
-        self.par_iter_mut().step_by(2)
-    }
-
-    /// Parallel iterator over parity pieces (odd indices)
-    #[inline]
-    pub fn par_parity(&self) -> impl IndexedParallelIterator<Item = &'_ PieceArray> + '_ {
-        self.par_iter().skip(1).step_by(2)
-    }
-
-    /// Mutable parallel iterator over parity pieces (odd indices)
-    #[inline]
-    pub fn par_parity_mut(
-        &mut self,
-    ) -> impl IndexedParallelIterator<Item = &'_ mut PieceArray> + '_ {
-        self.par_iter_mut().skip(1).step_by(2)
+        unsafe { Vec::from_raw_parts(value.as_mut_ptr(), PieceArray::SIZE, PieceArray::SIZE) }
     }
 }
