+++ subspace/crates/subspace-core-primitives/src/tests.rs
@@ -1,5 +1,5 @@
 use crate::pieces::PieceIndex;
-use crate::segments::{ArchivedHistorySegment, RecordedHistorySegment, SegmentIndex};
+use crate::segments::{RecordedHistorySegment, SegmentIndex};
 use crate::U256;
 
 #[test]
@@ -58,10 +58,10 @@ fn source_piece_index_conversion() {
         let segment_index = SegmentIndex::new(segment_index);
         let next_source_piece_index = PieceIndex::new(next_source_piece_index);
 
-        println!(
-            "{:?} {:?} {:?} {:?} {:?}",
-            piece_index, piece_position, source_position, segment_index, next_source_piece_index
-        );
+        // println!(
+        //     "{:?} {:?} {:?} {:?} {:?}",
+        //     piece_index, piece_position, source_position, segment_index, next_source_piece_index
+        // );
 
         assert_eq!(piece_index.position(), piece_position);
 
@@ -103,14 +103,14 @@ fn parity_piece_index_conversion() {
         let piece_index = PieceIndex::new(piece_index);
         let segment_index = SegmentIndex::new(segment_index);
 
-        println!("{:?} {:?} {:?}", piece_index, piece_position, segment_index,);
+        // println!("{:?} {:?} {:?}", piece_index, piece_position, segment_index);
 
         assert_eq!(piece_index.position(), piece_position);
 
         assert_eq!(piece_index.segment_index(), segment_index);
         assert!(!piece_index.is_source(), "{:?}", piece_index);
 
-        if piece_position as usize == ArchivedHistorySegment::NUM_PIECES - 1 {
+        if piece_position as usize == RecordedHistorySegment::NUM_PIECES - 1 {
             assert_eq!(segment_index.last_piece_index(), piece_index);
         }
 
@@ -135,10 +135,10 @@ fn parity_piece_index_conversion() {
 #[test]
 #[should_panic]
 fn parity_piece_index_position_panic() {
-    for &(piece_index, piece_position, segment_index) in PARITY_PIECE_INDEX_TEST_CASES {
+    for &(piece_index, _piece_position, _segment_index) in PARITY_PIECE_INDEX_TEST_CASES {
         let piece_index = PieceIndex::new(piece_index);
 
-        println!("{:?} {:?} {:?}", piece_index, piece_position, segment_index);
+        // println!("{:?} {:?} {:?}", piece_index, piece_position, segment_index);
 
         // Always panics
         piece_index.source_position();
@@ -148,10 +148,10 @@ fn parity_piece_index_position_panic() {
 #[test]
 #[should_panic]
 fn parity_piece_index_next_source_panic() {
-    for &(piece_index, piece_position, segment_index) in PARITY_PIECE_INDEX_TEST_CASES {
+    for &(piece_index, _piece_position, _segment_index) in PARITY_PIECE_INDEX_TEST_CASES {
         let piece_index = PieceIndex::new(piece_index);
 
-        println!("{:?} {:?} {:?}", piece_index, piece_position, segment_index);
+        // println!("{:?} {:?} {:?}", piece_index, piece_position, segment_index);
 
         // Always panics
         piece_index.next_source_index();
