+++ subspace/crates/subspace-erasure-coding/Cargo.toml
@@ -17,7 +17,7 @@ bench = false
 [dependencies]
 kzg.workspace = true
 rust-kzg-blst.workspace = true
-subspace-core-primitives.workspace = true
+subspace-core-primitives = { workspace = true, features = ["alloc"] }
 subspace-kzg.workspace = true
 
 [dev-dependencies]
